refactor(gui): move config related components into their own folder
