+++ src/gui/CMakeLists.txt
@@ -12,9 +12,9 @@ target_sources(${TARGET_NAME}
     PRIVATE 
         appearance/LookAndFeel.cpp
         components/MainComponent.cpp 
-        components/TorqueMapComponent.cpp
+        components/config/TorqueMapComponent.cpp
+        components/config/InverterConfigComponent.cpp
         components/TabbedComponent.cpp
-        components/InverterConfigComponent.cpp
         menubar/MenuBar.cpp
         windows/AboutWindow.cpp
         windows/MainWindow.cpp
