+++ src/gui/components/config/InverterConfigComponent.cpp
