+++ src/gui/components/config/TorqueMapComponent.cpp
@@ -6,8 +6,8 @@
 
 #include "TorqueMapComponent.h"
 
-#include "../../Application.h"
-#include "../utility/clip.h"
+#include "../../../Application.h"
+#include "../../utility/clip.h"
 
 namespace gui
 {
